refactor: finished all comment request from last pr
