+++ package.json
@@ -8,6 +8,7 @@
     "@testing-library/jest-dom": "^5.11.10",
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^12.8.3",
+    "date-fns": "^2.21.2",
     "firebase": "^8.4.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
