+++ src/App.js
@@ -1,8 +1,12 @@
 import { BrowserRouter } from "react-router-dom";
 import { Container } from "@material-ui/core";
-import { createMuiTheme, ThemeProvider } from "@material-ui/core/styles";
+import {
+  createMuiTheme,
+  ThemeProvider,
+  makeStyles,
+} from "@material-ui/core/styles";
 
-import useStyles from "./assets/styles/jss/styles";
+import PixBackground from "./assets/images/pix-background.png";
 import StyledAppBar from "./components/StyledAppBar";
 import Routes from "./routes/Routes";
 
@@ -11,6 +15,19 @@ const theme = createMuiTheme({
     fontFamily: "Original Surfer",
   },
 });
+
+const useStyles = makeStyles((theme) => ({
+  container: {
+    padding: theme.spacing(4, 2, 4),
+    maxWidth: "700px",
+    backgroundImage: `url(${PixBackground})`,
+    backgroundRepeat: "no-repeat",
+    backgroundPosition: "center center",
+    backgroundSize: "cover",
+    backgroundAttachment: "fixed",
+    height: "100%",
+  },
+}));
 
 const App = () => {
   const classes = useStyles();
